+++ lib/boost.php
@@ -1535,7 +1535,7 @@ function boost_rrdtool_function_create($local_data_id, $show_source, &$rrdtool_p
                             $spath .= '/' . $path;
 
                             $powner_id = fileowner($config['rra_path'] . $spath);
-                            $pgroup_id = fileowner($config['rra_path'] . $spath);
+                            $pgroup_id = filegroup($config['rra_path'] . $spath);
 
                             if ($powner_id != $owner_id) {
                                 $success = chown($config['rra_path'] . $spath, $owner_id);